Dif.limma

Differently Expression Analysis for genomic data. We apply limma package to conduct the analysis.

Usage

Dif.limma(
  Tumor_Data,
  Normal_Data,
  group = NULL,
  topk = NULL,
  sort.by = "p",
  adjust.method = "BH",
  RNAseq = FALSE
)

Arguments

Tumor_Data

A matrix representing the genomic data of cancer samples such as gene expression data, miRNA expression data.
For the matrix, the rows represent the genomic features, and the columns represent the cancer samples.

Normal_Data

A matrix representing the genomic data of Normal samples.
For the matrix, the rows represent the genomic features corresponding to the Tumor_Data, and the columns represent the normal samples.

group

A vector representing the subtype of each tumor sample in the Tumor_Data. The length of group is equal to the column number of Tumor_Data.

topk

The top number of different expression features that we want to extract in the return result.

sort.by

This is a parmeter of "topTable() in limma pacakge". "Character string specifying statistic to rank genes by. Possible values for topTable and toptable are "logFC", "AveExpr", "t", "P", "p", "B" or "none". (Permitted synonyms are "M" for "logFC", "A" or "Amean" for "AveExpr", "T" for "t" and "p" for "P".) Possibilities for topTableF are "F" or "none". Possibilities for topTreat are as for topTable except for "B"."

adjust.method

This is a parmeter of "topTable() in limma pacakge". Refer to the "method used to adjust the p-values for multiple testing. Options, in increasing conservatism, include "none", "BH", "BY" and "holm". See p.adjust for the complete list of options. A NULL value will result in the default adjustment method, which is "BH"."

RNAseq

A bool type representing the input datatype is a RNASeq or not. Default is FALSE for microarray data.

Value

A list representing the differently expression for each subtype comparing to the Normal group.

References

Smyth, Gordon K. "Limma: linear models for microarray data." Bioinformatics and computational biology solutions using R and Bioconductor. Springer New York, 2005. 397-420. Xu T, Le TD, Liu L, Su N, Wang R, Sun B, Colaprico A, Bontempi G, Li J. CancerSubtypes: an R/Bioconductor package for molecular cancer subtype identification, validation and visualization. Bioinformatics. 2017 Oct 1;33(19):3131-3133. doi: 10.1093/bioinformatics/btx378. PMID: 28605519.

Examples

data("mRNAexp")
mRNAexp=as.matrix(mRNAexp)
Normal_Data=as.matrix(mRNAexp[,sample(1:100,20)])
result=Dif.limma(Tumor_Data=mRNAexp,Normal_Data=Normal_Data,group=NULL,topk=NULL,RNAseq=FALSE)